refactor/#6: 패스 매핑 추가
