+++ sample-community/src/components/index.ts
@@ -1,2 +1,4 @@
-export * from "./Button";
-export * from "./TextField";
+export * from './Button';
+export * from './TextField';
+export * from './Post';
+export * from './PostPreView';
